+++ tests/Pest.php
@@ -1,5 +1,0 @@
-<?php
-
-use TranquilTools\FormBuilder\Tests\TestCase;
-
-uses(TestCase::class)->in(__DIR__);